% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidelta.R
\name{ei.delta}
\alias{ei.delta}
\title{Computes confidence intervals for interaction measures using the delta method by Hosmer and Lemeshow}
\usage{
ei.delta(model, coef = c(2, 3), CI.level = 0.95, type = "e")
}
\arguments{
\item{model}{A fitted model object of class glm or coxph}

\item{coef}{A vector of two numbers indicating the position of the two interaction terms in the model summary. The default is 2 and 3, which assumes that the two exposures are listed first before any additional covariates}

\item{CI.level}{A number indicating the confidence level for the CI estimation. The default is 0.95 CI}

\item{type}{A letter 'e' for effect modification or 'i' for interaction. The default is 'e'.}
}
\value{
A dataframe containing all the neccessary estimates required to full report effect modification or interaction with the CI estimates for interaction measures computed using the delta method
}
\description{
Computes confidence intervals for interaction measures using the delta method by Hosmer and Lemeshow
}
\examples{
ei.delta(model.glm, coef = c(2,3), CI.level = 0.95, type = 'i')
}
