% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDiscdata.R
\docType{data}
\name{HDiscdata}
\alias{HDiscdata}
\title{Case-control data of the effect of sports participation and smoking on herniated lumbar disc examined by Assmann et al.}
\format{
A data frame with 458 observations and 3 variables:
\describe{
  \item{h}{presence of herniated lumbar disc, the outcome}
  \item{ns}{presence or lack of sports participation, the first exposure}
  \item{smk}{smoking, the second exposure}
  ...
}
}
\source{
\doi{10.1097/00001648-199605000-00012}
}
\usage{
HDiscdata
}
\description{
A dataset containing three dichotomous variables: herniated lumbar disc, sports participation, and smoking
}
\keyword{datasets}
