% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionR.R
\name{interactionR}
\alias{interactionR}
\title{Full reporting of effect modification and interaction analysis}
\usage{
interactionR(
  model,
  exposure_names = c(),
  ci.type = "delta",
  ci.level = 0.95,
  em = T,
  recode = F
)
}
\arguments{
\item{model}{A fitted model object of class glm, clogit or coxph}

\item{exposure_names}{A character vector of two named binary exposure variables present in the fitted model: the default is an empty vector. If effect modification is being assessed, to get the right orientation of the table output, the first variable should be the putative effect modifier, the second, the main exposure. If it's interaction, the order doesn't matter.}

\item{ci.type}{A character string ("delta" or "mover") specifying the method to use for the estimation of CI for the measures of additive interaction. Default is "delta".}

\item{ci.level}{Magnitude of the returned CI level. Default is 0.95}

\item{em}{TRUE (the default), for effect modification assessment. FALSE, for interaction.}

\item{recode}{If TRUE, recodes the exposures - if at least one of the exposures is protective - such that the stratum with the lowest risk becomes the new reference category when the two exposures are considered jointly.}
}
\value{
Depending on if em was assessed or interaction, the function saves a publication-ready microsoft word Table corresponding to Table 1 or Table 3 respectively in Knol and Vanderweele (2012) to the working directory(see \url{https://doi.org/10.1093/ije/dyr218}).
It also returns an object of class "list" that includes:
#' \itemize{
  \item \code{Table} An object of class huxtable corresponding to the table saved in the working directory @seealso \code{\link[huxtable]{huxtable}}
  \item \code{dframe} A dataframe listing all the effect measures used to populate the table
}
}
\description{
For two binary exposures included in a logistic regression model as an interaction term for a binary outcome, \code{interactionR} returns a publication-ready table that include all effect measures neccessary to fully report effect modification or interaction analysis as recommended by Knol and Vanderweele (2012). Estimation of confidence intervals (CI) for measures of additive interaction is based on the delta method described by Hosmer and Lemeshow (1992) or the variance recovery 'mover' method described by Zou (2008).
}
\examples{
library (em.interaction)
data (OCdata) ## Case-control data from Rothman and Keller (1972) evaluating the joint effect of alcohol and smoking
## on oral cancer risk is included in the package (cited in Hosmer and Lemeshow (1992) and Zou (2008))

## fit the interaction model
model.glm <- glm(oc ~ alc*smk, family = binomial(link = "logit"), data = OCdata)

## Then pass the fitted model to the function
interactionR(model.glm, exposure_names = c("alc", "smk"), ci.type = "delta", ci.level = 0.95, em = F, recode = F)


}
