% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionR_bstrap.R
\name{interactionR_bstrap}
\alias{interactionR_bstrap}
\title{Computes confidence intervals for interaction measures using bootstrapping as described by Assmann et al (1996).}
\usage{
interactionR_bstrap(dat, seed = 5000)
}
\arguments{
\item{dat}{a data frame containing the variables required to estimate the measures of additive interactions in the following order of columns:
\itemize{
  \item \code{1)} the first column should be the outcome variable
  \item \code{2 & 3)} the second and third column should be the two exposures under study
  \item \code{others} the rest of the columns could be the covariates that requires adjustment in the model, if any.
}}

\item{seed}{The random seed to use for generating the bootstrap samples for confidence intervals (for reproducibility). Default is 5000, but can be set to any number.}
}
\value{
a dataframe containing effect estimates of additive interaction measures with bootstrapped 95% CI limits.
}
\description{
Computes confidence intervals for interaction measures using bootstrapping as described by Assmann et al (1996).
}
\examples{
data(HDiscdata) # loads the dataset from assmann et al. which is available in the package
head(HDiscdata, 2L) # prints the first two line to confirm the data is in the right format
interactionR_bstrap(HDiscdata)

}
\references{
Assmann SF, Hosmer DW, Lemeshow S, Mundt KA. Confidence intervals for measures of interaction. Epidemiology 1996:286-90.
}
