% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionR_delta.R
\name{interactionR_delta}
\alias{interactionR_delta}
\title{Computes confidence intervals for interaction measures using the delta method by Hosmer and Lemeshow}
\usage{
interactionR_delta(
  model,
  exposure_names = c(),
  ci.level = 0.95,
  em = T,
  recode = F
)
}
\arguments{
\item{model}{A fitted model object of class glm, clogit or coxph}

\item{exposure_names}{A character vector of two named binary exposure variables present in the fitted model: the default is an empty vector. If effect modification is being assessed, to get the right orientation of the table output, the first variable should be the putative effect modifier, the second, the main exposure. If it's interaction, the order doesn't matter.}

\item{ci.level}{Magnitude of the returned CI level. Default is 0.95}

\item{em}{TRUE (the default), for effect modification assessment. FALSE, for interaction.}

\item{recode}{If TRUE, recodes the exposures - if at least one of the exposures is protective - such that the stratum with the lowest risk becomes the new reference category when the two exposures are considered jointly.}
}
\value{
Depending on if em was assessed or interaction, the function saves a publication-ready microsoft word Table corresponding to Table 1 or Table 3 respectively in Knol and Vanderweele (2012) to the working directory(see \url{https://doi.org/10.1093/ije/dyr218}) and CI for interaction measures estimated with the delta method.
It also returns an object of class "list" that includes:
 \itemize{
  \item \code{Table} An object of class huxtable corresponding to the table saved in the working directory @seealso \code{\link[huxtable]{huxtable}}
  \item \code{dframe} A dataframe listing all the effect measures used to populate the table
}
}
\description{
Computes confidence intervals for interaction measures using the delta method by Hosmer and Lemeshow
}
\examples{
ei.delta(model.glm, coef = c(2,3), CI.level = 0.95, type = 'i')
}
