% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCdata.R
\docType{data}
\name{OCdata}
\alias{OCdata}
\title{Case-control study of the joint effect of alcohol and smoking on oral cancer by Rothman and Keller.}
\format{
A data frame with 458 observations and 3 variables:
\describe{
  \item{oc}{oral cancer, the outcome}
  \item{alc}{alcohol use, the first exposure}
  \item{smk}{smoking, the second exposure}
  ...
}
}
\source{
\url{https://doi.org/10.1016/0021-9681(72)90006-9}
}
\usage{
OCdata
}
\description{
A dataset containing three dichotomous variables: oral cancer, alcohol and smoking
}
\keyword{datasets}
